+++ src/LieGroups.jl
@@ -36,9 +36,20 @@ include("group_actions/group_operation_action.jl")
 include("groups/power_group.jl")
 include("groups/product_group.jl")
 include("groups/semidirect_product_group.jl")
+
 # Lie groups
+
 include("groups/translation_group.jl")
 include("groups/general_linear_group.jl")
+
+# includes generic implementations for O(n), U(n), SO(n), SO(n), so we load this first
+include("groups/unitary_group.jl")
+include("groups/orthogonal_group.jl")
+include("groups/special_unitary_group.jl")
+include("groups/special_orthogonal_group.jl")
+
+# Products of Groups
+include("groups/special_euclidean_group.jl")
 
 export LieGroup, LieAlgebra
 export PowerLieGroup, ProductLieGroup
@@ -61,12 +72,17 @@ export LeftSemidirectProductGroupOperation, RightSemidirectProductGroupOperation
 export AbstractGroupActionType
 export AbstractLeftGroupActionType, AbstractRightGroupActionType
 export LeftGroupOperationAction, RightGroupOperationAction
-export InverseLeftGroupOperationAction, InverseRightGroupOperationAction
 export GroupAction, GroupOperationAction
+export InverseLeftGroupOperationAction, InverseRightGroupOperationAction
+
 #
 #
 # Specific groups
-export TranslationGroup, GeneralLinearGroup
+export GeneralLinearGroup
+export OrthogonalGroup
+export SpecialEuclideanGroup, SpecialOrthogonalGroup, SpecialUnitaryGroup
+export TranslationGroup
+export UnitaryGroup
 
 export adjoint, adjoint!, apply, apply!
 export base_lie_group, base_manifold